backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.3
feedparser==6.0.11
pytest==8.3.4
pytest-asyncio==0.24.0
